+++ examples/src/main/java/io/dapr/examples/DaprMainApplication.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-package io.dapr.springboot;
+package io.dapr.examples;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
