+++ examples/src/main/java/io/dapr/examples/invoke/grpc/HelloWorldService.java
@@ -8,7 +8,6 @@ package io.dapr.examples.invoke.grpc;
 import com.google.protobuf.Any;
 import io.dapr.v1.AppCallbackGrpc;
 import io.dapr.v1.CommonProtos;
-import io.dapr.v1.DaprAppCallbackProtos;
 import io.grpc.Server;
 import io.grpc.ServerBuilder;
 import io.grpc.stub.StreamObserver;
