+++ examples/src/main/java/io/dapr/examples/invoke/http/DemoService.java
@@ -5,7 +5,7 @@
 
 package io.dapr.examples.invoke.http;
 
-import io.dapr.springboot.DaprApplication;
+import io.dapr.examples.DaprApplication;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.DefaultParser;
