+++ examples/src/main/java/io/dapr/examples/secrets/SecretClient.java
@@ -9,7 +9,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.dapr.client.DaprClient;
 import io.dapr.client.DaprClientBuilder;
 
-import java.io.IOException;
 import java.util.Map;
 
 /**
