+++ pom.xml
@@ -271,7 +271,6 @@
     <module>sdk-actors</module>
     <module>sdk-springboot</module>
     <module>examples</module>
-    <module>dapr-spring-boot-starter</module>
   </modules>
 
 </project>
