+++ sdk-springboot/src/main/java/io/dapr/springboot/DaprBeanPostProcessor.java
@@ -7,11 +7,8 @@ package io.dapr.springboot;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.dapr.Topic;
-import io.dapr.client.ObjectSerializer;
-import io.dapr.serializer.DefaultObjectSerializer;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.stereotype.Component;
