+++ sdk-springboot/src/main/resources/META-INF/spring.factories
